refactor: separate category input item
